Added explicit dtype to observation_space
To suppress warning coming from line 31 of spaces.Box

+++ gym/envs/classic_control/cartpole.py
@@ -38,7 +38,7 @@ class CartPoleEnv(gym.Env):
             np.finfo(np.float32).max])
 
         self.action_space = spaces.Discrete(2)
-        self.observation_space = spaces.Box(-high, high)
+        self.observation_space = spaces.Box(-high, high, dtype=np.float32)
 
         self.seed()
         self.viewer = None
